Put public message interface into msg.h and moved internal functions to msg.int.h.
