+++ BUILD
@@ -59,7 +59,7 @@ cc_library(
         "upb/decode.int.h",
         "upb/encode.c",
         "upb/msg.c",
-        "upb/msg.h",
+        "upb/msg.int.h",
         "upb/table.c",
         "upb/table.int.h",
         "upb/upb.c",
@@ -68,6 +68,7 @@ cc_library(
     hdrs = [
         "upb/decode.h",
         "upb/encode.h",
+        "upb/msg.h",
         "upb/upb.h",
         "upb/upb.hpp",
     ],
@@ -87,6 +88,7 @@ cc_library(
         "upb/decode_fast.c",
         "upb/decode_fast.h",
         "upb/msg.h",
+        "upb/msg.int.h",
         "upb/upb.int.h",
     ],
     copts = UPB_DEFAULT_COPTS,
