+++ upbc/protoc-gen-upb.cc
@@ -487,7 +487,7 @@ void WriteHeader(const protobuf::FileDescriptor* file, Output& output) {
   output(
       "#ifndef $0_UPB_H_\n"
       "#define $0_UPB_H_\n\n"
-      "#include \"upb/msg.h\"\n"
+      "#include \"upb/msg.int.h\"\n"
       "#include \"upb/decode.h\"\n"
       "#include \"upb/decode_fast.h\"\n"
       "#include \"upb/encode.h\"\n\n",
@@ -825,7 +825,7 @@ void WriteSource(const protobuf::FileDescriptor* file, Output& output,
 
   output(
       "#include <stddef.h>\n"
-      "#include \"upb/msg.h\"\n"
+      "#include \"upb/msg.int.h\"\n"
       "#include \"$0\"\n",
       HeaderFilename(file->name()));
